refactor brb components
