+++ brb/attr_input.py
@@ -1,0 +1,15 @@
+from typing import Dict, Any
+
+
+class AttributeInput():
+    """An input to the BRB system.
+
+    Consists of a set of antecedent attribute values and degrees of belief.
+
+    Attributes:
+        attr_input: X. Relates antecedent attributes with values and belief
+        degrees. Must follow same order of reference values as in the model.
+    """
+
+    def __init__(self, attr_input: Dict[str, Dict[Any, float]]):
+        self.attr_input = attr_input
